requirements.txt
backoff==1.10.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
feedparser==6.0.8
idna==2.10
opencage==1.2.2
Pillow==8.2.0
pycparser==2.20
pyOpenSSL==20.0.1
requests==2.25.1
sgmllib3k==1.0.0
six==1.16.0
urllib3==1.26.5
